Save generated loadout to URL and allow loading it from there
Loadout is stored as ?loadout=... query parameter. Reloading the page
will load the loadout from that if possible, meaning the URL can be
copied and shared with friends for them to see the same loadout. If the
value cannot be parsed correctly (e.g. someone has copied only some part
of it), a new loadout will be generated instead. Pressing the Go button
will always generate a new loadout.